Add command that creates breakpoint file that is later consumed by the launchRequest step
